requirements.txt
Django==3.0.7
djangorestframework==3.11.0
gunicorn==20.0.4
django-heroku==0.3.1
django-filter==2.2.0
mysqlclient==1.4.6
django-cors-headers==3.2.1
coreapi==2.3.3
djoser==2.0.3
flake8>=3.7.9
flake8-import-order>=0.18.1
flake8-print>=3.1.4
flake8-builtins>=1.5.2
flake8-bugbear>=20.1.4
flake8-return>=1.1.1